pyproject.toml
[build-system]
requires = ["hatchling", "hatch-vcs"]
build-backend = "hatchling.build"

[project]
name = "sqlean-driver"
dynamic = ["version"]
description = 'SQLAlchemy dialect for the sqlean.py SQLite wrapper'
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [{ name = "Edgar Ramírez Mondragón", email = "edgarrm358@gmail.com" }]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.7",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "sqlalchemy>=1.4.0",
  "sqlean-py>=0.21.5.1",
]

[project.entry-points."sqlalchemy.dialects"]
"sqlite.sqlean" = "sqlean_driver:SQLeanDialect"

[project.optional-dependencies]
dev = [
  "pytest",
]

[project.urls]
Documentation = "https://github.com/edgarrmondragon/sqlean-driver#readme"
Issues = "https://github.com/edgarrmondragon/sqlean-driver/issues"
Source = "https://github.com/edgarrmondragon/sqlean-driver"

[tool.hatch.version]
source = "vcs"

[tool.hatch.envs.default]
dependencies = ["coverage[toml]>=6.5", "pytest"]
[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"
test-cov = "coverage run -m pytest {args:tests}"
cov-xml = "coverage xml"
cov-report = ["- coverage combine", "coverage report"]
cov = ["test-cov", "cov-report"]

[[tool.hatch.envs.all.matrix]]
python = ["3.8", "3.9", "3.10", "3.11"]

[tool.hatch.envs.lint]
detached = true
dependencies = ["black>=23.1.0", "mypy>=1.0.0", "ruff>=0.0.243"]
[tool.hatch.envs.lint.scripts]
typing = "mypy --install-types --non-interactive {args:src/sqlean_driver tests}"
style = ["ruff check {args:.}", "black --check --diff {args:.}"]
style-gh = ["ruff check {args:.} --format github", "black --check --diff {args:.}"]
fmt = ["black {args:.}", "ruff --fix {args:.}", "style"]
all = ["style", "typing"]

[tool.black]
target-version = ["py38"]
line-length = 100
skip-string-normalization = true

[tool.pytest]
addopts = ["-vv"]

[tool.ruff]
line-length = 100
select = [
  "ALL",
]
target-version = "py38"

[tool.ruff.isort]
known-first-party = ["sqlean_driver"]
required-imports = ["from __future__ import annotations"]

[tool.ruff.flake8-tidy-imports]
ban-relative-imports = "all"

[tool.ruff.per-file-ignores]
# Tests can use magic values, assertions, and relative imports
"tests/**/*" = ["PLR2004", "S101", "TID252", "ANN201"]

[tool.ruff.pydocstyle]
convention = "google"

[tool.coverage.run]
source_pkgs = ["sqlean_driver", "tests"]
branch = true
parallel = true
omit = ["src/sqlean_driver/__about__.py"]

[tool.coverage.paths]
sqlean_driver = ["src/sqlean_driver", "*/sqlean-driver/src/sqlean_driver"]
tests = ["tests", "*/sqlean-driver/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov", "if __name__ == .__main__.:",
  "if t.TYPE_CHECKING:",
]
